# backend/document/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/v1/sync/sftp": {
      "post": {
        "tags": [
          "sftp-sync-controller"
        ],
        "operationId": "trigger",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/signatures/{applicationId}/review": {
      "post": {
        "tags": [
          "signature-controller"
        ],
        "operationId": "review",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/signatures/{applicationId}/request-approval": {
      "post": {
        "tags": [
          "signature-controller"
        ],
        "operationId": "requestApproval",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/signatures/upload": {
      "post": {
        "tags": [
          "signature-controller"
        ],
        "operationId": "uploadSignature",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignatureUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ai/signature-check": {
      "post": {
        "tags": [
          "ai-controller"
        ],
        "operationId": "checkSimilarity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "$ref": "#/components/schemas/AIResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/test": {
      "get": {
        "tags": [
          "hello-controller"
        ],
        "operationId": "hello",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/signatures": {
      "get": {
        "tags": [
          "signature-controller"
        ],
        "operationId": "getAllSignatures",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignatureResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/signatures/{applicationId}": {
      "get": {
        "tags": [
          "signature-controller"
        ],
        "operationId": "getSignatureDetails",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "/": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SignatureUploadRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "imageBase64": {
            "type": "string"
          }
        }
      },
      "AIRequest": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AIResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "SignatureResponse": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "signatureImage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}