# backend/src/main/java/com/demo/dto/AIRequest.java
package com.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AIRequest {
    private UUID applicationId;
}