# backend/src/main/java/com/demo/dto/AIResponse.java
package com.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AIResponse {
    private String result;
    private Double score;
}